Completed JS restructure, added progress bars for server execution
This commit completed the JS restructure and adds progress bars so the user can see the server-side execution.
