+++ config/requirements.txt
@@ -1,3 +1,4 @@
+alive-progress==3.1.5
 numpy==1.26.1
 scipy==1.11.3
 scikit-learn==1.3.2
@@ -8,3 +9,4 @@ dash==2.14.2
 flask==3.0.0
 flask-caching==2.1.0
 flask-cors==4.0.0
+regex==2023.10.3
