+++ demo.py
@@ -1,3 +1,11 @@
+"""
+Author:     Blake McBride (blakepm2@illinois.edu)
+Created:    12/02/2023
+
+Overview:   This file contains a demo script to open sample webpages and start running the server so the user can demo Historian
+"""
+
+# standard imports
 import subprocess 
 
 # open the sample webpages
