+++ extension/index.html
@@ -23,6 +23,10 @@
     <iframe id='graph' frameborder="0"></iframe>  
   </div>
   <!-- <script src="getPlot.js"></script> -->
-  <script src="test.js"></script>
+  <script src="src/spinners.js"></script>
+  <script src="src/fetchHistory.js"></script>
+  <script src="src/sendURLStoServer.js"></script>
+  <script src="src/getGraph.js"></script>
+  <script src="script.js"></script>
 </body>
 </html>
