+++ extension/src/spinners.js
@@ -1,0 +1,7 @@
+function showSpinner() {
+    document.getElementById('spinner-wrapper').style.display = 'flex';
+}
+
+function hideSpinner() {
+    document.getElementById('spinner-wrapper').style.display = 'none';
+}
